Rename getEC2Regions file
