+++ src/utils/AWS/EC2/getEC2Regions.ts
